Created seperate component for course
